requirements.txt
streamlit
pypdf
langchain
faiss-cpu
langchain-groq
python-dotenv
requests
beautifulsoup4
faiss-cpu
langchain-community
sentence-transformers
matplotlib
langchain-text-splitters